+++ src/moveit_controller/src/servo_keyboard.cpp
@@ -1,83 +1,23 @@
-#include <signal.h>
-#include <stdio.h>
-#include <termios.h>
-#include <unistd.h>
+#include "servo_keyboard.h"
 
-#include <control_msgs/msg/joint_jog.hpp>
-#include <geometry_msgs/msg/twist_stamped.hpp>
-#include <rclcpp/rclcpp.hpp>
-
-// Define used keys
-#define KEYCODE_RIGHT 0x43
-#define KEYCODE_LEFT 0x44
-#define KEYCODE_UP 0x41
-#define KEYCODE_DOWN 0x42
-#define KEYCODE_PERIOD 0x2E
-#define KEYCODE_SEMICOLON 0x3B
-#define KEYCODE_1 0x31
-#define KEYCODE_2 0x32
-#define KEYCODE_3 0x33
-#define KEYCODE_4 0x34
-#define KEYCODE_5 0x35
-#define KEYCODE_6 0x36
-#define KEYCODE_7 0x37
-#define KEYCODE_Q 0x71
-#define KEYCODE_W 0x77
-#define KEYCODE_E 0x65
-#define KEYCODE_R 0x72
-
-// Some constants used in the Servo Teleop demo
-const std::string TWIST_TOPIC = "/servo_node/delta_twist_cmds";
-const std::string JOINT_TOPIC = "/servo_node/delta_joint_cmds";
-const size_t ROS_QUEUE_SIZE = 10;
-const std::string EEF_FRAME_ID = "Link_7";
-const std::string BASE_FRAME_ID = "Link_1";
-
-// A class for reading the key inputs from the terminal
-class KeyboardReader {
- public:
-  KeyboardReader() : kfd(0) {
-    // get the console in raw mode
-    tcgetattr(kfd, &cooked);
-    struct termios raw;
-    memcpy(&raw, &cooked, sizeof(struct termios));
-    raw.c_lflag &= ~(ICANON | ECHO);
-    // Setting a new line, then end of file
-    raw.c_cc[VEOL] = 1;
-    raw.c_cc[VEOF] = 2;
-    tcsetattr(kfd, TCSANOW, &raw);
+KeyboardReader::KeyboardReader() : kfd(0) {
+  // get the console in raw mode
+  tcgetattr(kfd, &cooked);
+  struct termios raw;
+  memcpy(&raw, &cooked, sizeof(struct termios));
+  raw.c_lflag &= ~(ICANON | ECHO);
+  // Setting a new line, then end of file
+  raw.c_cc[VEOL] = 1;
+  raw.c_cc[VEOF] = 2;
+  tcsetattr(kfd, TCSANOW, &raw);
+}
+void KeyboardReader::readOne(char* c) {
+  int rc = read(kfd, c, 1);
+  if (rc < 0) {
+    throw std::runtime_error("read failed");
   }
-  void readOne(char* c) {
-    int rc = read(kfd, c, 1);
-    if (rc < 0) {
-      throw std::runtime_error("read failed");
-    }
-  }
-  void shutdown() { tcsetattr(kfd, TCSANOW, &cooked); }
-
- private:
-  int kfd;
-  struct termios cooked;
-};
-
-// Converts key-presses to Twist or Jog commands for Servo, in lieu of a
-// controller
-class KeyboardServo {
- public:
-  KeyboardServo();
-  int keyLoop();
-
- private:
-  void spin();
-
-  rclcpp::Node::SharedPtr nh_;
-
-  rclcpp::Publisher<geometry_msgs::msg::TwistStamped>::SharedPtr twist_pub_;
-  rclcpp::Publisher<control_msgs::msg::JointJog>::SharedPtr joint_pub_;
-
-  std::string frame_to_publish_;
-  double joint_vel_cmd_;
-};
+}
+void KeyboardReader::shutdown() { tcsetattr(kfd, TCSANOW, &cooked); }
 
 KeyboardServo::KeyboardServo()
     : frame_to_publish_(BASE_FRAME_ID), joint_vel_cmd_(1.0) {
@@ -87,28 +27,6 @@ KeyboardServo::KeyboardServo()
       TWIST_TOPIC, ROS_QUEUE_SIZE);
   joint_pub_ = nh_->create_publisher<control_msgs::msg::JointJog>(
       JOINT_TOPIC, ROS_QUEUE_SIZE);
-}
-
-KeyboardReader input;
-
-void quit(int sig) {
-  (void)sig;
-  input.shutdown();
-  rclcpp::shutdown();
-  exit(0);
-}
-
-int main(int argc, char** argv) {
-  rclcpp::init(argc, argv);
-  KeyboardServo keyboard_servo;
-
-  signal(SIGINT, quit);
-
-  int rc = keyboard_servo.keyLoop();
-  input.shutdown();
-  rclcpp::shutdown();
-
-  return rc;
 }
 
 void KeyboardServo::spin() {
@@ -250,4 +168,24 @@ int KeyboardServo::keyLoop() {
   }
 
   return 0;
+}
+
+void quit(int sig) {
+  (void)sig;
+  input.shutdown();
+  rclcpp::shutdown();
+  exit(0);
+}
+
+int main(int argc, char** argv) {
+  rclcpp::init(argc, argv);
+  KeyboardServo keyboard_servo;
+
+  signal(SIGINT, quit);
+
+  int rc = keyboard_servo.keyLoop();
+  input.shutdown();
+  rclcpp::shutdown();
+
+  return rc;
 }
